py2/backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==41.0.7
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
PyMySQL==1.1.0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
